+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/datagen/datapack/ModEnchantmentProvider.kt
@@ -2,20 +2,17 @@ package dev.aaronhowser.mods.geneticsresequenced.datagen.datapack
 
 import dev.aaronhowser.mods.geneticsresequenced.GeneticsResequenced
 import dev.aaronhowser.mods.geneticsresequenced.datagen.tag.ModItemTagsProvider
-import dev.aaronhowser.mods.geneticsresequenced.util.OtherUtil
-import net.minecraft.core.RegistrySetBuilder
 import net.minecraft.core.registries.Registries
 import net.minecraft.data.worldgen.BootstrapContext
 import net.minecraft.resources.ResourceKey
 import net.minecraft.world.entity.EquipmentSlotGroup
 import net.minecraft.world.item.enchantment.Enchantment
 
-object ModEnchantmentProvider : RegistrySetBuilder() {
+object ModEnchantmentProvider {
 
 	val DELICATE_TOUCH = createRk("delicate_touch")
 
 	fun bootstrap(context: BootstrapContext<Enchantment>) {
-
 		val itemGetter = context.lookup(Registries.ITEM)
 
 		context.register(
@@ -32,7 +29,6 @@ object ModEnchantmentProvider : RegistrySetBuilder() {
 				)
 			).build(DELICATE_TOUCH.location())
 		)
-
 	}
 
 	private fun createRk(name: String): ResourceKey<Enchantment> =