+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/datagen/tag/ModDamageTypeTagsProvider.kt
@@ -1,14 +1,11 @@
 package dev.aaronhowser.mods.geneticsresequenced.datagen.tag
 
 import dev.aaronhowser.mods.geneticsresequenced.GeneticsResequenced
-import dev.aaronhowser.mods.geneticsresequenced.util.OtherUtil
+import dev.aaronhowser.mods.geneticsresequenced.datagen.datapack.ModDamageTypeProvider
 import net.minecraft.core.HolderLookup
-import net.minecraft.core.registries.Registries
 import net.minecraft.data.PackOutput
 import net.minecraft.data.tags.DamageTypeTagsProvider
-import net.minecraft.resources.ResourceKey
 import net.minecraft.tags.DamageTypeTags
-import net.minecraft.world.damagesource.DamageType
 import net.neoforged.neoforge.common.data.ExistingFileHelper
 import java.util.concurrent.CompletableFuture
 
@@ -21,32 +18,24 @@ class ModDamageTypeTagsProvider(
 	override fun addTags(provider: HolderLookup.Provider) {
 		this.tag(DamageTypeTags.NO_IMPACT)
 			.add(
-				STEP_ON_SYRINGE,
-				USE_SYRINGE,
-				USE_SCRAPER
+				ModDamageTypeProvider.STEP_ON_SYRINGE,
+				ModDamageTypeProvider.USE_SYRINGE,
+				ModDamageTypeProvider.USE_SCRAPER
 			)
 
 		this.tag(DamageTypeTags.NO_ANGER)
 			.add(
-				STEP_ON_SYRINGE,
-				USE_SYRINGE,
-				USE_SCRAPER
+				ModDamageTypeProvider.STEP_ON_SYRINGE,
+				ModDamageTypeProvider.USE_SYRINGE,
+				ModDamageTypeProvider.USE_SCRAPER
 			)
 
 		this.tag(DamageTypeTags.NO_KNOCKBACK)
 			.add(
-				STEP_ON_SYRINGE,
-				USE_SYRINGE,
-				USE_SCRAPER
+				ModDamageTypeProvider.STEP_ON_SYRINGE,
+				ModDamageTypeProvider.USE_SYRINGE,
+				ModDamageTypeProvider.USE_SCRAPER
 			)
-	}
-
-	companion object {
-		private fun createType(name: String): ResourceKey<DamageType> = ResourceKey.create(Registries.DAMAGE_TYPE, GeneticsResequenced.modResource(name))
-
-		val STEP_ON_SYRINGE: ResourceKey<DamageType> = createType("step_on_syringe")
-		val USE_SYRINGE: ResourceKey<DamageType> = createType("use_syringe")
-		val USE_SCRAPER: ResourceKey<DamageType> = createType("use_scraper")
 	}
 
 }